File and data restructure: easier to work with summarized coverage
